ga4-params Move new code to TS file version
